Moving uploader from components to containers
